+++ scandeval/benchmark.py
@@ -16,10 +16,8 @@ from .angry_tweets import AngryTweetsBenchmark
 from .twitter_sent import TwitterSentBenchmark
 from .twitter_subj import TwitterSubjBenchmark
 from .europarl_subj import EuroparlSubjBenchmark
-from .europarl1 import Europarl1Benchmark
-from .europarl2 import Europarl2Benchmark
-from .lcc1 import Lcc1Benchmark
-from .lcc2 import Lcc2Benchmark
+from .europarl_sent import EuroparlSentBenchmark
+from .lcc import LccBenchmark
 from .dkhate import DkHateBenchmark
 from .utils import InvalidBenchmark
 
@@ -99,12 +97,10 @@ class Benchmark:
             ('angry-tweets', 'Angry Tweets', AngryTweetsBenchmark(**params)),
             ('twitter-sent', 'TwitterSent', TwitterSentBenchmark(**params)),
             ('twitter-subj', 'TwitterSubj', TwitterSubjBenchmark(**params)),
+            ('europarl-sent', 'EuroparlSent', EuroparlSentBenchmark(**params)),
             ('europarl-subj', 'EuroparlSubj', EuroparlSubjBenchmark(**params)),
             ('dkhate', 'DKHate', DkHateBenchmark(**params)),
-            ('europarl1', 'Europarl1', Europarl1Benchmark(**params)),
-            ('europarl2', 'Europarl2', Europarl2Benchmark(**params)),
-            ('lcc1', 'LCC1', Lcc1Benchmark(**params)),
-            ('lcc2', 'LCC2', Lcc2Benchmark(**params))
+            ('lcc', 'LCC', LccBenchmark(**params))
         ]
 
     @staticmethod