+++ scandeval/datasets.py
@@ -134,6 +134,8 @@ def load_dkhate() -> Tuple[dict, dict, dict, dict]:
 def load_twitter_subj() -> Tuple[dict, dict, dict, dict]:
     '''Load the TwitterSubj dataset.
 
+    This dataset is the subjectivity part of the TwitterSent dataset.
+
     Returns:
         tuple:
             Four dicts, `X_train`, `X_test`, `y_train` and `y_test`, where
@@ -153,6 +155,8 @@ def load_twitter_subj() -> Tuple[dict, dict, dict, dict]:
 def load_europarl_subj() -> Tuple[dict, dict, dict, dict]:
     '''Load the EuroparlSubj dataset.
 
+    This dataset is the subjectivity part of the Europarl2 dataset.
+
     Returns:
         tuple:
             Four dicts, `X_train`, `X_test`, `y_train` and `y_test`, where
@@ -169,8 +173,10 @@ def load_europarl_subj() -> Tuple[dict, dict, dict, dict]:
     return X_train, X_test, y_train, y_test
 
 
-def load_europarl1() -> Tuple[dict, dict, dict, dict]:
-    '''Load the Europarl1 dataset.
+def load_europarl_sent() -> Tuple[dict, dict, dict, dict]:
+    '''Load the EuroparlSent dataset.
+
+    This dataset is the sentiment part of the Europarl2 dataset.
 
     Returns:
         tuple:
@@ -180,7 +186,7 @@ def load_europarl1() -> Tuple[dict, dict, dict, dict]:
             contains the target vectors.
     '''
     base_url = ('https://raw.githubusercontent.com/saattrupdan/ScandEval/'
-                'main/datasets/europarl1/')
+                'main/datasets/europarl_sent/')
     train_url = base_url + 'train.jsonl'
     test_url = base_url + 'test.jsonl'
     X_train, y_train = _get_dataset_from_url(train_url, 'text', 'label')
@@ -188,8 +194,10 @@ def load_europarl1() -> Tuple[dict, dict, dict, dict]:
     return X_train, X_test, y_train, y_test
 
 
-def load_europarl2() -> Tuple[dict, dict, dict, dict]:
-    '''Load the Europarl2 dataset.
+def load_lcc() -> Tuple[dict, dict, dict, dict]:
+    '''Load the LCC dataset.
+
+    This dataset is the concatenation of the LCC1 and LCC2 datasets.
 
     Returns:
         tuple:
@@ -199,45 +207,7 @@ def load_europarl2() -> Tuple[dict, dict, dict, dict]:
             contains the target vectors.
     '''
     base_url = ('https://raw.githubusercontent.com/saattrupdan/ScandEval/'
-                'main/datasets/europarl2/')
-    train_url = base_url + 'train.jsonl'
-    test_url = base_url + 'test.jsonl'
-    X_train, y_train = _get_dataset_from_url(train_url, 'text', 'label')
-    X_test, y_test = _get_dataset_from_url(test_url, 'text', 'label')
-    return X_train, X_test, y_train, y_test
-
-
-def load_lcc1() -> Tuple[dict, dict, dict, dict]:
-    '''Load the LCC1 dataset.
-
-    Returns:
-        tuple:
-            Four dicts, `X_train`, `X_test`, `y_train` and `y_test`, where
-            `X_train` and `X_test` corresponds to the feature matrices for the
-            training and test split, respectively, and `y_train` and `y_test`
-            contains the target vectors.
-    '''
-    base_url = ('https://raw.githubusercontent.com/saattrupdan/ScandEval/'
-                'main/datasets/lcc1/')
-    train_url = base_url + 'train.jsonl'
-    test_url = base_url + 'test.jsonl'
-    X_train, y_train = _get_dataset_from_url(train_url, 'text', 'label')
-    X_test, y_test = _get_dataset_from_url(test_url, 'text', 'label')
-    return X_train, X_test, y_train, y_test
-
-
-def load_lcc2() -> Tuple[dict, dict, dict, dict]:
-    '''Load the LCC2 dataset.
-
-    Returns:
-        tuple:
-            Four dicts, `X_train`, `X_test`, `y_train` and `y_test`, where
-            `X_train` and `X_test` corresponds to the feature matrices for the
-            training and test split, respectively, and `y_train` and `y_test`
-            contains the target vectors.
-    '''
-    base_url = ('https://raw.githubusercontent.com/saattrupdan/ScandEval/'
-                'main/datasets/lcc2/')
+                'main/datasets/lcc/')
     train_url = base_url + 'train.jsonl'
     test_url = base_url + 'test.jsonl'
     X_train, y_train = _get_dataset_from_url(train_url, 'text', 'label')