+++ scandeval/lcc.py
@@ -1,4 +1,4 @@
-'''Sentiment evaluation of a language model on the Europarl1 dataset'''
+'''Sentiment evaluation of a language model on the LCC dataset'''
 
 from datasets import Dataset
 from typing import Tuple
@@ -6,14 +6,14 @@ import logging
 
 from .sentiment_classification import SentimentClassificationBenchmark
 from .utils import doc_inherit
-from .datasets import load_europarl1
+from .datasets import load_lcc
 
 
 logger = logging.getLogger(__name__)
 
 
-class Europarl1Benchmark(SentimentClassificationBenchmark):
-    '''Benchmark of language models on the Europarl1 dataset.
+class LccBenchmark(SentimentClassificationBenchmark):
+    '''Benchmark of language models on the LCC dataset.
 
     Args:
         cache_dir (str, optional):
@@ -44,14 +44,14 @@ class Europarl1Benchmark(SentimentClassificationBenchmark):
                  cache_dir: str = '.benchmark_models',
                  evaluate_train: bool = False,
                  verbose: bool = False):
-        super().__init__(name='Europarl1',
+        super().__init__(name='LCC',
                          cache_dir=cache_dir,
                          evaluate_train=evaluate_train,
                          verbose=verbose)
 
     @doc_inherit
     def _load_data(self) -> Tuple[Dataset, Dataset]:
-        X_train, X_test, y_train, y_test = load_europarl1()
+        X_train, X_test, y_train, y_test = load_lcc()
         train_dict = dict(doc=X_train['text'],
                           orig_label=y_train['label'])
         test_dict = dict(doc=X_test['text'],